+++ app/Http/Controllers/PermissionController.php
@@ -18,16 +18,28 @@ class PermissionController extends Controller
 {
     public function index()
     {
+        if(!Auth::user()->can('view_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('permission.index');
     }
 
     public function create()
     {
+        if(!Auth::user()->can('create_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('permission.create');
     }
 
     public function store(StorePermissionRequest $request)
     {
+        if(!Auth::user()->can('create_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         Permission::create($request->only('name'));
 
         return redirect()->route('permission.index')->with('create', 'Permission is Successfully Create');
@@ -35,10 +47,18 @@ class PermissionController extends Controller
 
     public function edit(Permission $permission)
     {
+        if(!Auth::user()->can('edit_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('permission.edit', compact('permission'));
     }
 
     public function update($id, UpdatePermissionRequest $request) {
+        if(!Auth::user()->can('edit_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $permission = Permission::findOrFail($id);
         $permission->update($request->only('name'));
 
@@ -47,12 +67,20 @@ class PermissionController extends Controller
 
     public function destroy(Permission $permission) 
     {
+        if(!Auth::user()->can('delete_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $permission->delete();
         return 'success';
     }
 
     public function ssd()
     {
+        if(!Auth::user()->can('view_permission')) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $permissions = Permission::query();
         return datatables($permissions)
             ->addColumn('plus-icon', function($each) {
@@ -64,8 +92,17 @@ class PermissionController extends Controller
                     Carbon::parse($each->updated_at)->format('H:i:s A');
             })
             ->addColumn('action', function($each) {
-                $edit_icon = '<a href="'.route('permission.edit', $each->id).'" class="text-warning"><i class="fas fa-user-edit"></i></a>';
-                $delete_icon = '<a href="#" class="text-danger delete_btn" data-id="'.$each->id.'"><i class="fas fa-trash"></i></a>';
+                $edit_icon = '';
+                $delete_icon = '';
+
+                if(Auth::user()->can('edit_permission')) {
+                    $edit_icon = '<a href="'.route('permission.edit', $each->id).'" class="text-warning"><i class="fas fa-user-edit"></i></a>';
+                }
+
+                if(Auth::user()->can('delete_permission')) {
+                    $delete_icon = '<a href="#" class="text-danger delete_btn" data-id="'.$each->id.'"><i class="fas fa-trash"></i></a>';
+                }
+
                 return '<div class="action_icon">'. $edit_icon . $delete_icon .'</div>';
             })
             ->rawColumns(['action'])
